create ChartInfo, PricesChart and VolumeChart components
